Moving external libraries not easily available through npm to own folder
