+++ web/src/main/webapp/js/lib/leaflet_numbered_markers.js
